Add Constants.swift to watchOS project
