+++ OCKWatchSample Extension/Constants.swift
@@ -1,0 +1,17 @@
+//
+//  Constants.swift
+//  OCKWatchSample Extension
+//
+//  Created by Corey Baker on 12/10/20.
+//  Copyright © 2020 Apple. All rights reserved.
+//
+
+import Foundation
+
+enum Constants {
+    static let group = "group.netrecon.ParseCarekitSample"
+    static let parseUserKey = "requestParseUser"
+    static let parseRemoteClockIDKey = "requestRemoteClockID"
+    static let requestSync = "requestSync"
+    static let userLoggedIn = "userLoggedIn"
+}
